Add case where fishes only have a chance to spawn.
fishesCount in FishingSpotScriptableObject is now a float where 0.1 == 10% spawn rate

+++ Assets/Scripts/ScriptableObjects/FishingSpotScriptableObject.cs
@@ -7,5 +7,5 @@ public class FishingSpotScriptableObject : ScriptableObject
 {
     [Header("Fishes and FishesCount must have the same number of elements")]
     public List<ScriptableObject> fishes;
-    public List<int> fishesCount;
+    public List<float> fishesCount;
 }
